Refactor VeteransCrisisLine to be self-contained across multiple components
